// src/constants/forms/types.ts
export enum FormTypes {
  LOGIN = 'LOGIN',
  REGISTER = 'REGISTER',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD',
  RESET_PASSWORD = 'RESET_PASSWORD',
  PROFILE = 'PROFILE',
}

export enum FormFieldTypes {
  NAME_FIELD_TYPE = 'name',
  EMAIL_FIELD_TYPE = 'email',
  PASSWORD_FIELD_TYPE = 'password',
  REQUEST_FIELD_TYPE = 'request',
  CODE_FIELD_TYPE = 'token',
}